+++ packages/react/src/atomic/Icon/src/utils/gradient/cssParser/index.ts
@@ -1,1 +1,1 @@
-export { parseLinearGradient } from "./cssParser"; 
+export { parseLinearGradient } from "./cssParser";